叫料需求，excel解析成需求列表

+++ wms-cloud/wms-api/wms-api-binIn/src/main/java/com/bosch/binin/api/domain/Stock.java
@@ -68,17 +68,17 @@ public class Stock extends BaseEntity {
     /**
      * 总库存
      */
-    private Integer totalStock;
+    private Double totalStock;
 
     /**
      * 冻结库存
      */
-    private Integer freezeStock;
+    private Double freezeStock;
 
     /**
      * 可用库存
      */
-    private Integer availableStock;
+    private Double availableStock;
 
     /**
      * PO号
